DCI/requirements.txt
Bio==1.4.0
interval==1.0.0
intervaltree==3.1.0
ipython==8.5.0
matplotlib==3.5.1
nbformat==5.7.0
numpy==1.21.5
pandas==1.4.2
Pillow==9.2.0
scikit_learn==1.1.2
torch==1.10.2
